requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.9.2
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.10=pypi_0
aiosignal=1.3.2=pypi_0
altair=5.5.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.7.0=pypi_0
asgiref=3.8.1=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
async-timeout=4.0.3=pypi_0
attrs=24.3.0=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.2.1=pypi_0
blinker=1.9.0=pypi_0
build=1.2.2.post1=pypi_0
bzip2=1.0.8=h2466b09_7
ca-certificates=2024.12.14=h56e8100_0
cachetools=5.5.0=pypi_0
certifi=2024.12.14=pypi_0
charset-normalizer=3.4.0=pypi_0
chroma-hnswlib=0.7.3=pypi_0
chromadb=0.5.3=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pyhd8ed1ab_1
coloredlogs=15.0.1=pypi_0
comm=0.2.2=pyhd8ed1ab_1
cpython=3.9.21=py39hd8ed1ab_1
debugpy=1.8.11=py39ha51f57c_0
decorator=5.1.1=pyhd8ed1ab_1
deprecated=1.2.15=pypi_0
durationpy=0.9=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
fastapi=0.115.6=pypi_0
filelock=3.16.1=pypi_0
flatbuffers=24.3.25=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.10.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
google-ai-generativelanguage=0.6.10=pypi_0
google-api-core=2.24.0=pypi_0
google-api-python-client=2.155.0=pypi_0
google-auth=2.37.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-generativeai=0.8.3=pypi_0
googleapis-common-protos=1.66.0=pypi_0
greenlet=3.1.1=pypi_0
grpcio=1.68.1=pypi_0
grpcio-status=1.68.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httplib2=0.22.0=pypi_0
httptools=0.6.4=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.27.0=pypi_0
humanfriendly=10.0=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=pyha770c72_1
importlib-resources=6.4.5=pypi_0
ipykernel=6.29.5=pyh4bbf305_0
ipython=8.18.1=pyh7428d3b_3
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh5737063_1
krb5=1.21.3=hdf4eb48_0
kubernetes=31.0.0=pypi_0
langchain=0.3.12=pypi_0
langchain-core=0.3.25=pypi_0
langchain-text-splitters=0.3.3=pypi_0
langsmith=0.2.3=pypi_0
libffi=3.4.2=h8ffe710_5
liblzma=5.6.3=h2466b09_1
libsodium=1.0.20=hc70643c_0
libsqlite=3.47.2=h67fdade_0
libzlib=1.3.1=h2466b09_2
lxml=5.3.0=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mdurl=0.1.2=pypi_0
mmh3=5.0.1=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
narwhals=1.18.4=pypi_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.2.1=pypi_0
numpy=1.26.4=pypi_0
oauthlib=3.2.2=pypi_0
onnxruntime=1.19.2=pypi_0
openssl=3.4.0=h2466b09_0
opentelemetry-api=1.29.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.29.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.29.0=pypi_0
opentelemetry-instrumentation=0.50b0=pypi_0
opentelemetry-instrumentation-asgi=0.50b0=pypi_0
opentelemetry-instrumentation-fastapi=0.50b0=pypi_0
opentelemetry-proto=1.29.0=pypi_0
opentelemetry-sdk=1.29.0=pypi_0
opentelemetry-semantic-conventions=0.50b0=pypi_0
opentelemetry-util-http=0.50b0=pypi_0
orjson=3.10.12=pypi_0
overrides=7.7.0=pypi_0
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=pypi_0
parso=0.8.4=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.0.0=pypi_0
pip=24.3.1=pyh8b19718_0
platformdirs=4.3.6=pyhd8ed1ab_1
posthog=3.7.4=pypi_0
prompt-toolkit=3.0.48=pyha770c72_1
propcache=0.2.1=pypi_0
proto-plus=1.25.0=pypi_0
protobuf=5.29.1=pypi_0
psutil=6.1.0=py39ha55e580_0
pure_eval=0.2.3=pyhd8ed1ab_1
pyarrow=18.1.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pydantic=2.10.3=pypi_0
pydantic-core=2.27.1=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.18.0=pyhd8ed1ab_1
pymupdf=1.25.1=pypi_0
pyparsing=3.2.0=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.2.0=pypi_0
pyreadline3=3.5.4=pypi_0
python=3.9.21=h37870fc_1_cpython
python-dateutil=2.9.0.post0=pyhff2d567_1
python-docx=1.1.2=pypi_0
python-dotenv=1.0.1=pypi_0
python_abi=3.9=5_cp39
pytz=2024.2=pypi_0
pywin32=307=py39ha51f57c_3
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=py39h03e5c00_3
referencing=0.35.1=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=13.9.4=pypi_0
rpds-py=0.22.3=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.5=pypi_0
scikit-learn=1.6.0=pypi_0
scipy=1.13.1=pypi_0
sentence-transformers=3.3.1=pypi_0
setuptools=75.6.0=pyhff2d567_1
shellingham=1.5.4=pypi_0
six=1.17.0=pyhd8ed1ab_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.36=pypi_0
stack_data=0.6.3=pyhd8ed1ab_1
starlette=0.41.3=pypi_0
streamlit=1.41.1=pypi_0
sympy=1.13.1=pypi_0
tenacity=9.0.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.13=h5226925_1
tokenizers=0.21.0=pypi_0
toml=0.10.2=pypi_0
tomli=2.2.1=pypi_0
torch=2.5.1=pypi_0
tornado=6.4.2=py39ha55e580_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.47.0=pypi_0
typer=0.15.1=pypi_0
typing_extensions=4.12.2=pyha770c72_1
tzdata=2024.2=pypi_0
ucrt=10.0.22621.0=h57928b3_1
uritemplate=4.1.1=pypi_0
urllib3=2.2.3=pypi_0
uvicorn=0.34.0=pypi_0
vc=14.3=ha32ba9b_23
vc14_runtime=14.42.34433=he29a5d6_23
vs2015_runtime=14.42.34433=hdffcdeb_23
watchdog=6.0.0=pypi_0
watchfiles=1.0.3=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
websocket-client=1.8.0=pypi_0
websockets=14.1=pypi_0
wheel=0.45.1=pyhd8ed1ab_1
wrapt=1.17.0=pypi_0
yarl=1.18.3=pypi_0
zeromq=4.3.5=ha9f60a1_7
zipp=3.21.0=pyhd8ed1ab_1
